rename decorators.py into deprecations.py
